feat: restructure component hierarchy
