+++ src/components/TaskManager/TaskList/TaskItem.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Task } from "../types/task";
+import { Task } from "../../../types/task";
 
 type Props = {
   task: Task;
